requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
torch==2.0.1
sentence-transformers==2.2.2
streamlit==1.24.1
langchain==0.0.234